requirements.txt
python-pptx~=1.0.2
marko~=2.2.2